Issue #1: handle broken gravity sensor.
If the device has a gravity sensor, but it's broken, then we get a valid
Sensor object which never reports any values to our listener.  If we
don't hear from the sensor, then we fall back on the same behavior as if
we don't have a sensor at all (use movement on the touch screen to
decide whether to call decideZeroFs()).